Simplify logic further by printing the matching part up front
Each time through the loop, we need to call merge for the matching
part no matter what happens.  Just get it out of the way up front and
then we no longer have to worry about whether the match part was an
actual character or if it was the beginning of the line.

Change-Id: Id3364a119e90dd85cc2ea0e3f2021cc9b3954ff1